Rename base classes
